requirements.txt
flask
pandas
matplotlib
seaborn
plotly
langchain
langchain-groq
python-dotenv
dateparser
